+++ R/input_checks.r
@@ -550,13 +550,11 @@ check_inputs_node_time_to_event <- function(data, parents, sim_time, name,
         length(body(prob_fun)) != 0)
 
     # check content of prob_fun_args
-    if (length(setdiff(names(formals(prob_fun)),
-                       c("data", "sim_time"))) != 0) {
-      for (i in seq_len(length(setdiff(names(formals(prob_fun)),
-                                 c("data", "sim_time"))))) {
-        if(!is.element(setdiff(names(formals(prob_fun)),
-                               c("data", "sim_time"))[i],
-                       names(prob_fun_args))) {
+    arg_names <- setdiff(names(formals(prob_fun)), c("data", "sim_time"))
+    if (length(args) != 0) {
+      for (i in seq_len(length(arg_names))) {
+        if(!arg_names[i] %in% names(prob_fun_args) &
+           class(formals(prob_fun)[[arg_names[i]]])=="name") {
           stop("All parameters of 'prob_fun' except 'data' and 'sim_time'",
                " must be included in the node_td() call if they don't have a",
                " default value.")