+++ R/node.r
@@ -223,5 +223,20 @@ print.DAG.node <- function(x, ...) {
 ## S3 summary method for DAG.node objects
 #' @export
 summary.DAG.node <- function(object, ...) {
-  print.DAG.node(x=object, ...)
+
+  if (!is.character(object[[1]])) {
+    str_equations <- character()
+    for (i in seq_len(length(object))) {
+      str_equations_i <- structural_equation(object[[i]])
+      str_equations <- c(str_equations, str_equations_i)
+    }
+  } else {
+    str_equations <- structural_equation(object)
+  }
+  str_equations_print <- align_str_equations(str_equations)
+
+  cat("A DAG.node object using the following structural equation(s):\n\n")
+  cat(str_equations_print, sep="\n")
+
+  return(invisible(str_equations))
 }