+++ tests/testthat/test_align_str_equations.r
@@ -1,0 +1,7 @@
+
+test_that("general test case", {
+  equations <- c("A ~ some stuff",
+                 "another_name ~ more stuff")
+  expect_equal(align_str_equations(equations),
+               c("           A ~ some stuff", "another_name ~ more stuff"))
+})