+++ tests/testthat/test_extract_function_name.r
@@ -1,0 +1,9 @@
+
+test_fun <- function(x) {
+  print(10 + x)
+}
+
+test_that("general test", {
+  out <- extract_function_name(test_fun)
+  expect_equal(out, "test_fun")
+})